+++ lib/my_animated_icon.dart
@@ -1,3 +1,0 @@
-library;
-
-export 'src/my_animated_icon.dart';